update exoPlayer , post  and  comment   2020-08-19 10.36.AM

+++ app/src/main/java/com/gyobeom29/hipboard/PostInfo.java
@@ -98,11 +98,11 @@ public class PostInfo implements Parcelable {
         this.publisher = publisher;
     }
 
-    public void setViews(int views) {
+    public void setViews(long views) {
         this.views = views;
     }
 
-    public void setLikeCount(int likeCount) {
+    public void setLikeCount(long likeCount) {
         this.likeCount = likeCount;
     }
